+++ pkg/servers/product_base/logic/product_category.go
@@ -2,6 +2,7 @@ package logic
 
 import (
 	"errors"
+	"fmt"
 
 	"github.com/CloudSilk/CloudSilk/pkg/model"
 	"github.com/CloudSilk/CloudSilk/pkg/proto"
@@ -74,5 +75,13 @@ func GetProductCategoryByIDs(ids []string) ([]*model.ProductCategory, error) {
 }
 
 func DeleteProductCategory(id string) (err error) {
+	var count int64
+	if err = model.DB.DB().Where(&model.ProductCategoryAttribute{ProductCategoryID: id}).Count(&count).Error; err != nil {
+		return
+	}
+	if count > 0 {
+		return fmt.Errorf("数据冲突，请先清空关联此特性的产品类别特性")
+	}
+
 	return model.DB.DB().Delete(&model.ProductCategory{}, "`id` = ?", id).Error
 }
